Allow :artifact field, #63 plus allow for re-processing. #62
Agents were updated to include an extra :artifact field but this wasn't reflected
in the Percolator, meaning that Evidence Records went un-processed. This commit
fixes #63.

Because inputs went un-processed, and because Percolator can now safely re-ingest
input, add a new feature, #62, to allow the Kafka group to be based on major.minor
version number, meaning that whenever a large change happens, the entire data will
be (safely) re-ingested.